refactor: rewrite directory structure
