+++ test/index.test.ts
@@ -1,5 +1,5 @@
 import test from 'tape'
-import api from '../api'
+import api from '../src/api'
 
 test('dummy test', (t) => {
   t.is(typeof api, 'function')
